calender should work now
git-svn-id: svn://141.76.65.164/kp-salespoint@136 eb53f5e8-9ec5-4bc0-8d13-06c033283886

+++ SalespointFramework/src/org/salespointframework/core/calendar/Calendar.java
@@ -13,26 +13,31 @@ import org.salespointframework.util.Filter;
 public interface Calendar<T extends CalendarEntry> {
     
     /**
-     * Return all Entries
-     * @return {@link Iterable} of all entries that match the filter criteria 
+     * Returns all Entries that match the filter criteria.
+     * If no entries were found an empty {@link Iterable} will be returned, not null;
+     * 
+     * 
+     * @return Iterable of all entries that match the filter criteria
+     * 
+     * @see {@link Filter}
      */
     Iterable<T> getEntries(Filter<T> filter);
     
     /**
-     * Return the calendar entry which has the given id or null if no entry exists with this id
+     * Returns the calendar entry which has the given id or null if no entry exists with this id.
      * @param id of the entry that should be returned
-     * @return the entry with given id or null
+     * @return the entry with the given id or null
      */
-    T getEntryByID(int id);
+    T getEntryByID(long id);
     
     /**
-     * Adds the given entry to the calendar 
+     * Adds the given entry to the calendar.
      * @param entry the entry that should be added to the calendar
      */
     void addEntry(T entry);
     
     /**
-     * Deletes the entry with the given id
+     * Deletes the entry with the given id.
      * @param id of the entry that should be removed from calendar
      */
     void deleteEntry(int id);
